requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
google-generativeai==0.3.2
plotly==5.18.0
matplotlib==3.8.3
seaborn==0.13.2
python-dotenv==1.0.1